refactoring the recursive logic to allow nested components and fixing some data inconsistency
