+++ frontend/src/common/PostCard.js
@@ -1,26 +1,24 @@
-import React from 'react';
+import React, { useState } from 'react';
 
 import AuthorBox from './AuthorBox';
 import PostAction from './PostAction';
 import VotingWidget from './VotingWidget';
 import { distanceDate } from '../utilities/formatDate';
 
-const PostCard = ({
-  onMessageIconClick,
-  post,
-  link,
-  isComment = false,
-  children,
-  ...others
-}) => {
+const PostCard = ({ post, link, isComment = false, children, ...others }) => {
   const {
     user,
     created,
     vote_rank,
     content,
+    comments,
     comment_count,
     share_count,
   } = post;
+
+  const [showComments, setShowComments] = useState(false);
+
+  const toggleComments = () => setShowComments(!showComments);
 
   return (
     <div className={`${isComment && 'post-card--comment'}`} {...others}>
@@ -41,13 +39,28 @@ const PostCard = ({
           {content}
         </div>
       </div>
-
       <PostAction
-        onMessageIconClick={onMessageIconClick}
+        onMessageIconClick={toggleComments}
         comments={comment_count}
         link={link}
         shares={share_count}
       />
+      {showComments && (
+        <div className="post-comments-wrapper">
+          {comments.map((comment) => (
+            <div key={comment.id} className="post-comment">
+              <PostCard post={comment} link={'/'} isComment={true}>
+                <div className="comment__mentioned">
+                  Replying to
+                  {comment.reply_at?.map((user) => (
+                    <span key={user.id}> @{user.username}</span>
+                  ))}
+                </div>
+              </PostCard>
+            </div>
+          ))}
+        </div>
+      )}{' '}
     </div>
   );
 };
